rename: filterをformatterに整理: `non_kanji_block_number.rs` -> `chome_with_arabic_numerals.rs`
