+++ core/src/formatter.rs
@@ -1,3 +1,4 @@
+pub(crate) mod chome_with_arabic_numerals;
 pub(crate) mod fullwidth_character;
 pub(crate) mod house_number;
 pub(crate) mod informal_town_name_notation;
