+++ core/src/parser/filter.rs
@@ -1,5 +1,3 @@
-pub mod non_kanji_block_number;
-
 pub trait Filter {
     fn apply(self, input: String) -> String;
 }
